Added admin prefix to each "critical"  & modified tournament index view
on the route file, I added an "admin" prefix before each critical URLs.

On the tournament index view, I hidded each admin buttons if you are not an connected administrator

+++ resources/views/tournament/index.blade.php
@@ -31,12 +31,17 @@
 								<div class="hour"> {{$tournament->start_time}} à {{$tournament->end_time}} </div>
 								<div class="date"> {{$tournament->start_date->format('d.m.Y')}} au {{$tournament->end_date->format('d.m.Y')}} </div>
 
-								<a href="{{route('tournaments.edit', $tournament->id)}}" title="Éditer le tournoi" class="edit"><i class="fa fa-lg fa-pencil action" aria-hidden="true"></i></a>
-								{{ Form::open(array('url' => route('tournaments.destroy', $tournament->id), 'method' => 'delete')) }}
-									<button type="button" class="button-delete" data-name="{{ $tournament->name }}" data-type="tournament">
-					                    <i class="fa fa-lg fa-trash-o action" aria-hidden="true"></i>
-					                </button>
-								{{ Form::close() }}
+								@if(Auth::check())
+									@if(Auth::user()->role == 'administrator')
+										<a href="{{route('tournaments.edit', $tournament->id)}}" title="Éditer le tournoi" class="edit"><i class="fa fa-lg fa-pencil action" aria-hidden="true"></i></a>
+										{{ Form::open(array('url' => route('tournaments.destroy', $tournament->id), 'method' => 'delete')) }}
+											<button type="button" class="button-delete" data-name="{{ $tournament->name }}" data-type="tournament">
+							                    <i class="fa fa-lg fa-trash-o action" aria-hidden="true"></i>
+							                </button>
+										{{ Form::close() }}
+									@endif
+								@endif
+
 							</div>
 
 						</div>
@@ -46,7 +51,11 @@
 
 		</div>
 
-		<a href="{{route('tournaments.create')}}" title="Créer un tournoi"><i class="fa fa-plus-circle fa-4x" aria-hidden="true"></i></a>
+		@if(Auth::check())
+			@if(Auth::user()->role == 'administrator')
+				<a href="{{route('tournaments.create')}}" title="Créer un tournoi"><i class="fa fa-plus-circle fa-4x" aria-hidden="true"></i></a>
+			@endif
+		@endif
 
 	</div>
 @stop
